requirements.txt
Flask==2.2.2
azure-identity==1.11.0
azure-storage-blob==12.14.1